int overflow fix

+++ mappedptr.h
@@ -77,7 +77,7 @@ uint64_t _align16(uint64_t ptrval) { if ( (ptrval & 15) != 0 ) ptrval += 16 - (p
 void *alloc_aligned_buffer(uint64_t allocsize)
 {
 	extern int posix_memalign (void **__memptr, size_t __alignment, size_t __size);
-	if ( allocsize > 192L*1024*1024*1024 )
+	if ( allocsize > ((uint64_t)192L)*1024*1024*1024 )
 		printf("%llu\n",(long long)allocsize),fatal("negative allocsize");
 	void *ptr;
 	allocsize = _align16(allocsize);
